refactor: Moved modifyRecipe to recipe tools
